+++ __tests__/spot/nft/nftWithdrawHistory.test.js
@@ -1,0 +1,20 @@
+/* global describe, it, expect */
+const { nockMock, buildQueryString, SpotClient } = require('../../testUtils/testSetup')
+const { mockResponse, startTime, endTime, limit, recvWindow } = require('../../testUtils/mockData')
+const parameters = {
+  startTime,
+  endTime,
+  limit,
+  recvWindow
+}
+
+describe('#nftWithdrawHistory', () => {
+  it('should fetch NFT withdraw history', () => {
+    nockMock(`/sapi/v1/nft/history/withdraw?${buildQueryString({ ...parameters })}`)(mockResponse)
+
+    return SpotClient.nftWithdrawHistory(parameters).then(response => {
+      expect(response).toBeDefined()
+      expect(response.data).toEqual(mockResponse)
+    })
+  })
+})