+++ examples/spot/pay/payHistory.js
@@ -2,16 +2,12 @@ const Spot = require('../../../src/spot')
 
 const apiKey = ''
 const apiSecret = ''
+
 const client = new Spot(apiKey, apiSecret)
 
-client.withdraw(
-  'BNB', // coin
-  'bnb_address', // withdraw address
-  1, // amount
-  { // optional parameters
-    network: 'BNB',
-    name: 'address name',
-    walletType: 0
+client.payHistory(
+  {
+    startTimestamp: 1637570276000
   }
 ).then(response => client.logger.log(response.data))
   .catch(error => client.logger.error(error))