+++ examples/spot/sub_account/subAccountApiAddIp.js
@@ -2,16 +2,12 @@ const Spot = require('../../../src/spot')
 
 const apiKey = ''
 const apiSecret = ''
+
 const client = new Spot(apiKey, apiSecret)
 
-client.withdraw(
-  'BNB', // coin
-  'bnb_address', // withdraw address
-  1, // amount
-  { // optional parameters
-    network: 'BNB',
-    name: 'address name',
-    walletType: 0
-  }
+client.subAccountApiAddIp(
+  'alice@test.com',
+  'subAccountApiKey',
+  '1.2.3.4'
 ).then(response => client.logger.log(response.data))
   .catch(error => client.logger.error(error))